+++ tests/test_nudatus.py
@@ -16,19 +16,19 @@ def test_get_version():
     Ensure a call to get_version returns the expected string.
     """
     result = nudatus.get_version()
-    assert result == '.'.join([str(i) for i in nudatus._VERSION])
+    assert result == ".".join([str(i) for i in nudatus._VERSION])
 
 
 def test_mangle_script():
     """
     Check that mangle returns the expected output
     """
-    script = ''
-    real_mangled = b''
-    with open('tests/bigscript.py') as f:
+    script = ""
+    real_mangled = b""
+    with open("tests/bigscript.py") as f:
         script = f.read()
     assert len(script) > 0
-    with open('tests/bigscript_mangled.py') as f:
+    with open("tests/bigscript_mangled.py") as f:
         real_mangled = f.read()
     assert len(real_mangled) > 0
     mangled = nudatus.mangle(script)
@@ -40,8 +40,8 @@ def test_mangle_with_bad_syntax():
     Check that mangle throws an
     exception for a badly formatted script
     """
-    script = ''
-    with open('tests/bigscript_bad.py') as f:
+    script = ""
+    with open("tests/bigscript_bad.py") as f:
         script = f.read()
     assert len(script) > 0
     with pytest.raises(tokenize.TokenError):
@@ -49,7 +49,7 @@ def test_mangle_with_bad_syntax():
 
 
 def test_main_without_file(capfd):
-    with mock.patch('sys.argv', ['nudatus']):
+    with mock.patch("sys.argv", ["nudatus"]):
         with pytest.raises(SystemExit) as ex:
             nudatus.main()
             assert ex.value.code == 1
@@ -59,11 +59,11 @@ def test_main_without_file(capfd):
 
 
 def test_main_with_file_without_output_file(capfd):
-    script = ''
-    with open('tests/bigscript_mangled.py') as f:
+    script = ""
+    with open("tests/bigscript_mangled.py") as f:
         script = f.read()
     assert len(script) > 0
-    with mock.patch('sys.argv', ['nudatus', 'tests/bigscript.py']):
+    with mock.patch("sys.argv", ["nudatus", "tests/bigscript.py"]):
         nudatus.main()
         out, err = capfd.readouterr()
         assert len(err) == 0
@@ -71,20 +71,19 @@ def test_main_with_file_without_output_file(capfd):
 
 
 def test_main_with_file_with_output_file(capfd):
-    expected = ''
-    with open('tests/bigscript_mangled.py') as f:
+    expected = ""
+    with open("tests/bigscript_mangled.py") as f:
         expected = f.read()
     assert len(expected) > 0
-    script = ''
-    with open('tests/bigscript.py') as f:
+    script = ""
+    with open("tests/bigscript.py") as f:
         script = f.read()
     assert len(script) > 0
-    with mock.patch('sys.argv',
-                    ['nudatus', 'tests/bigscript.py', 'testout.py']):
+    with mock.patch("sys.argv", ["nudatus", "tests/bigscript.py", "testout.py"]):
         m = mock.mock_open(read_data=script)
-        with mock.patch.object(builtins, 'open', m):
+        with mock.patch.object(builtins, "open", m):
             nudatus.main()
-        m.assert_called_with('testout.py', 'w')
+        m.assert_called_with("testout.py", "w")
         handle = m()
         handle.write.assert_called_with(expected)
         out, err = capfd.readouterr()
@@ -94,10 +93,10 @@ def test_main_with_file_with_output_file(capfd):
 
 def test_main_with_bad_script(capfd):
     with pytest.raises(SystemExit) as ex:
-        with mock.patch('sys.argv', ['nudatus', 'tests/bigscript_bad.py']):
+        with mock.patch("sys.argv", ["nudatus", "tests/bigscript_bad.py"]):
             nudatus.main()
         assert ex.value.code == 1
     out, err = capfd.readouterr()
     assert len(out) == 0
     assert len(err) > 0
-    assert err.startswith('Error mangling tests/bigscript_bad.py:')
+    assert err.startswith("Error mangling tests/bigscript_bad.py:")